url_shortener/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==1.7.1
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
django-prometheus==2.3.1
